+++ package.json
@@ -2,10 +2,10 @@
   "name": "reddit-market-bot",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "ts-node --esm ./src/index.ts"
+    "start": "ts-node --esm ./src/app.ts"
   },
   "author": "",
   "license": "ISC",
